fix(category): Category 컴포넌트 디렉토리 refactor 및 Faculty 페이지 내 Category 컴포넌트 적용 로직 수정 (#53)
