+++ public/data/faculty.json
@@ -1,128 +1,131 @@
-[
-  {
-    "id": "1",
-    "thumbnail": "lee-geun.jpg",
-    "info": {
-      "name": "Lee Geun",
-      "major": "Smart Mobility Design",
-      "category": "Smart Mobility"
+{
+  "majorList": ["All", "Smart Mobility", "Spatial", "Product"],
+  "faculty": [
+    {
+      "id": "1",
+      "thumbnail": "lee-geun.jpg",
+      "info": {
+        "name": "Lee Geun",
+        "major": "Smart Mobility Design",
+        "category": "Smart Mobility"
+      }
+    },
+    {
+      "id": "2",
+      "thumbnail": "koo-sang.jpg",
+      "info": {
+        "name": "Koo Sang",
+        "major": "Smart Mobility Design",
+        "category": "Smart Mobility"
+      }
+    },
+    {
+      "id": "3",
+      "thumbnail": "kim-ju-yeon.jpg",
+      "info": {
+        "name": "Kim Ju Yeon",
+        "major": "Spatial Design",
+        "category": "Spatial"
+      }
+    },
+    {
+      "id": "4",
+      "thumbnail": "lee-jeong-gyo.jpg",
+      "info": {
+        "name": "Lee Jeong Gyo",
+        "major": "Spatial Marketing Design",
+        "category": "Spatial"
+      }
+    },
+    {
+      "id": "5",
+      "thumbnail": "cho-taek-yeon.jpg",
+      "info": {
+        "name": "Cho Taek Yeon",
+        "major": "Spatial Design",
+        "category": "Spatial"
+      }
+    },
+    {
+      "id": "6",
+      "thumbnail": "choi-ik-seo.jpg",
+      "info": {
+        "name": "Choi Ik Seo",
+        "major": "Spatial Design",
+        "category": "Spatial"
+      }
+    },
+    {
+      "id": "7",
+      "thumbnail": "hwang-seong-gul.jpg",
+      "info": {
+        "name": "Hwang Seong Gul",
+        "major": "Industrial Design",
+        "category": "Product"
+      }
+    },
+    {
+      "id": "8",
+      "thumbnail": "kim-sook-yeon.jpg",
+      "info": {
+        "name": "Kim Sook Yeon",
+        "major": "Interaction/Service Design",
+        "category": "Interaction"
+      }
+    },
+    {
+      "id": "9",
+      "thumbnail": "kim-eun-young.jpg",
+      "info": {
+        "name": "Kim Eun Young",
+        "major": "Industrial Design",
+        "category": "Product"
+      }
+    },
+    {
+      "id": "10",
+      "thumbnail": "kim-jae-yeop.jpg",
+      "info": {
+        "name": "Kim Jae Yeop",
+        "major": "Inetraction/Industrial Design",
+        "category": "Interaction"
+      }
+    },
+    {
+      "id": "11",
+      "thumbnail": "eo-joon-hyuk.jpg",
+      "info": {
+        "name": "Eo Joon Hyuk",
+        "major": "Smart Mobility Design",
+        "category": "Smart Mobility"
+      }
+    },
+    {
+      "id": "12",
+      "thumbnail": "lee-gil-ok.jpg",
+      "info": {
+        "name": "Lee Gil Ok",
+        "major": "Industrial Design",
+        "category": "Product"
+      }
+    },
+    {
+      "id": "13",
+      "thumbnail": "lee-hee-jin.jpg",
+      "info": {
+        "name": "Lee Hee Jin",
+        "major": "Exhibition Design",
+        "category": "Product"
+      }
+    },
+    {
+      "id": "14",
+      "thumbnail": "noh-seung-il.jpg",
+      "info": {
+        "name": "Noh Seung Il",
+        "major": "Smart Mobility Design",
+        "category": "Smart Mobility"
+      }
     }
-  },
-  {
-    "id": "2",
-    "thumbnail": "koo-sang.jpg",
-    "info": {
-      "name": "Koo Sang",
-      "major": "Smart Mobility Design",
-      "category": "Smart Mobility"
-    }
-  },
-  {
-    "id": "3",
-    "thumbnail": "kim-ju-yeon.jpg",
-    "info": {
-      "name": "Kim Ju Yeon",
-      "major": "Spatial Design",
-      "category": "Spatial"
-    }
-  },
-  {
-    "id": "4",
-    "thumbnail": "lee-jeong-gyo.jpg",
-    "info": {
-      "name": "Lee Jeong Gyo",
-      "major": "Spatial Marketing Design",
-      "category": "Spatial"
-    }
-  },
-  {
-    "id": "5",
-    "thumbnail": "cho-taek-yeon.jpg",
-    "info": {
-      "name": "Cho Taek Yeon",
-      "major": "Spatial Design",
-      "category": "Spatial"
-    }
-  },
-  {
-    "id": "6",
-    "thumbnail": "choi-ik-seo.jpg",
-    "info": {
-      "name": "Choi Ik Seo",
-      "major": "Spatial Design",
-      "category": "Spatial"
-    }
-  },
-  {
-    "id": "7",
-    "thumbnail": "hwang-seong-gul.jpg",
-    "info": {
-      "name": "Hwang Seong Gul",
-      "major": "Industrial Design",
-      "category": "Product"
-    }
-  },
-  {
-    "id": "8",
-    "thumbnail": "kim-sook-yeon.jpg",
-    "info": {
-      "name": "Kim Sook Yeon",
-      "major": "Interaction/Service Design",
-      "category": "Interaction"
-    }
-  },
-  {
-    "id": "9",
-    "thumbnail": "kim-eun-young.jpg",
-    "info": {
-      "name": "Kim Eun Young",
-      "major": "Industrial Design",
-      "category": "Product"
-    }
-  },
-  {
-    "id": "10",
-    "thumbnail": "kim-jae-yeop.jpg",
-    "info": {
-      "name": "Kim Jae Yeop",
-      "major": "Inetraction/Industrial Design",
-      "category": "Interaction"
-    }
-  },
-  {
-    "id": "11",
-    "thumbnail": "eo-joon-hyuk.jpg",
-    "info": {
-      "name": "Eo Joon Hyuk",
-      "major": "Smart Mobility Design",
-      "category": "Smart Mobility"
-    }
-  },
-  {
-    "id": "12",
-    "thumbnail": "lee-gil-ok.jpg",
-    "info": {
-      "name": "Lee Gil Ok",
-      "major": "Industrial Design",
-      "category": "Product"
-    }
-  },
-  {
-    "id": "13",
-    "thumbnail": "lee-hee-jin.jpg",
-    "info": {
-      "name": "Lee Hee Jin",
-      "major": "Exhibition Design",
-      "category": "Product"
-    }
-  },
-  {
-    "id": "14",
-    "thumbnail": "noh-seung-il.jpg",
-    "info": {
-      "name": "Noh Seung Il",
-      "major": "Smart Mobility Design",
-      "category": "Smart Mobility"
-    }
-  }
-]
+  ]
+}
